pyproject.toml
[project]
name = "gimmegit"
version = "0.4.0.dev0"
description = "A tool for cloning GitHub repos and creating branches."
authors = [
    { name = "Dave Wilding", email = "tech@dpw.me" },
]
license = "MIT"
license-files = ["LICENSE"]
requires-python = ">= 3.10"
dependencies = [
    "gitpython>=3.1,<4",
    "pre-commit>=4.4,<5",
    "PyGithub>=2.6,<3",
]

[project.scripts]
gimmegit = "gimmegit._cli:main"

[dependency-groups]
dev = [
    "ruff",
    "pytest",
    "ty>=0.0.1a33",
]

[build-system]
requires = [
    "uv_build>=0.7.21,<0.8",
]
build-backend = "uv_build"

[tool.ruff]
line-length = 99

[tool.ty.environment]
# Help ty resolve local imports in our testing code.
root = [
    "./src",
    "./tests/functional",
    "./tests/unit"
]
